[IMPROVE] extension 리팩토링 (#40)
* feat: 상세 저장 혹은 빠른 저장 중일 때 로딩 토스트 띄워주도록 구현

* feat: 웹 페이지 바로 갈 수 있도록 버튼 추가

* feat: 상세 저장 페이지에 있는 태그, 카테고리 선택 컴포넌트 분리

* feat: 요약 성공 시 카테고리와 태그 prefetchQuery 적용
