+++ apps/extension/src/lib/tanstack/query/category.ts
@@ -7,5 +7,6 @@ export const useGetCategories = () => {
     queryKey: [CATEGORY.GET_CATEGORIES],
     queryFn: getCategories,
     select: (data) => data.result.map((category) => category.name),
+    staleTime: 1000 * 60,
   });
 };
