--- apps/extension/src/pages/redirect.tsx
@@ -1,3 +1,0 @@
-export default function Redirect() {
-  return <div>Redirect</div>;
-}
